pyproject.toml
[project]
name = "prompti"
version = "0.8.15"
description = "Provider-agnostic asynchronous prompt engine"
authors = [
  { name = "wolvever" }, { name = "gaoxiaoyang02" }
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2",
  "jinja2>=3",
  "async-lru",
  "httpx[http2]>=0.25",
  "tenacity>=8",
  "aiofiles",
  "pyyaml",
  "prometheus-client",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "xxhash",
  "python-dotenv>=1.1.0",
  "semantic-version>=2.10.0",
  "pytest-asyncio",
]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio"]
litellm = [
    "litellm>=1.73.1",
]

[tool.uv]
# uv is used for dependency management during development

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
extend-exclude = ["tests/data"]
ignore = ["PYI001"]  # Ignore untyped import rule

[tool.ruff.lint]
select = [
    "D",   # pycodestyle docstring
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "S",   # bandit
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "B018", "D"]
"examples/*" = ["B018"]

[tool.mypy]
ignore_missing_imports = true

[tool.pyright]
exclude = ["**/__pycache__"]
reportMissingImports = "warning"
reportMissingModuleSource = "none"

[tool.pyright.strict]
